PDFAgent/requirements.txt
phidata
python-dotenv
packaging
pypdf
fastapi
uvicorn
groq
sqlalchemy
pgvector
psycopg2-binary
